requirements.txt
streamlit==1.45.1
pandas==2.2.3
matplotlib==3.10.0
numpy==2.1.3
lxml==5.3.0
beautifulsoup4==4.12.3
requests==2.32.3
json5==0.9.25
Pillow==11.1.0
